Parsers for embedded subtitles and external subtitles are handled separately.
