+++ lib_ass_media/src/main/java/io/github/peerless2012/ass/media/common/AssC.kt
@@ -1,0 +1,6 @@
+package io.github.peerless2012.ass.media.common
+
+/**
+ * Indicates the track is external subtitles.
+ */
+public const val ROLE_FLAG_EXTERNAL_SUBTITLES = 1 shl 16
